Improve performance (#19)
* Replace double with float

* Improve README.md

* Improve displaying of README.md on github

+++ Imagine.Models/ColorHsv.cs
@@ -1,22 +1,22 @@
 namespace Imagine.Models;
 
-public readonly record struct ColorHsv(double H, double S, double V)
+public readonly record struct ColorHsv(float H, float S, float V)
 {
 	public static explicit operator Color(ColorHsv value)
 	{
-		const double yellow = 1D;
-		const double green = 2D;
-		const double cyan = 3D;
-		const double blue = 4D;
-		const double magenta = 5D;
-		const double red = 6D;
+		const float yellow = 1F;
+		const float green = 2F;
+		const float cyan = 3F;
+		const float blue = 4F;
+		const float magenta = 5F;
+		const float red = 6F;
 
-		var h = red * value.H.Modulo(1D);
-		var s = double.Clamp(value.S, 0D, 1D);
-		var v = double.Clamp(value.V, 0D, 1D);
+		var h = red * value.H.Modulo(1F);
+		var s = float.Clamp(value.S, 0F, 1F);
+		var v = float.Clamp(value.V, 0F, 1F);
 
-		var min = v * (1D - s);
-		var between = v * (1D - (s * double.Abs((h % 2D) - 1D)));
+		var min = v * (1F - s);
+		var between = v * (1F - (s * float.Abs((h % 2F) - 1F)));
 		var max = v;
 
 		return h switch
