+++ Imagine.Models/Intercept.cs
@@ -1,11 +1,11 @@
 namespace Imagine.Models;
 
 public readonly record struct Intercept(
-	double Distance,
+	float Distance,
 	Vector3 Normal,
 	Color Color)
 {
-	public Intercept(double distance, Vector3 normal)
+	public Intercept(float distance, Vector3 normal)
 		: this(
 			  Distance: distance,
 			  Normal: normal,
