+++ Imagine.Models/Line3.cs
@@ -2,6 +2,6 @@ namespace Imagine.Models;
 
 public readonly record struct Line3(Vector3 Origin, Vector3 Direction)
 {
-	public Vector3 At(double distance) =>
+	public Vector3 At(float distance) =>
 		Origin + (Direction * distance);
 }
